[FIX] #593 game, match 재수정
