+++ gg-pingpong-data/src/main/java/gg/pingpong/data/game/redis/RankRedis.java
@@ -46,16 +46,16 @@ public class RankRedis implements Serializable {
 		this.statusMessage = msg;
 	}
 
-	public static RankRedis from(UserDto user, Integer ppp, String tierImageUri) {
+	public static RankRedis from(Long id, String intraId, String textColor, Integer ppp, String tierImageUri) {
 		RankRedis rankRedis = RankRedis.builder()
-			.userId(user.getId())
-			.intraId(user.getIntraId())
+			.userId(id)
+			.intraId(intraId)
 			.ppp(ppp)
 			.wins(0)
 			.losses(0)
 			.statusMessage("")
 			.tierImageUri(tierImageUri)
-			.textColor(user.getTextColor())
+			.textColor(textColor)
 			.build();
 		return rankRedis;
 	}
